Make a fake manifest for bind mount
To disable c2.ffmpeg completely, we will disable the service and not
include the manifest of the hidl. In order to do that we will make a
fake xml file that is nearly empty. If we turn on ffmpeg codec we will
bind mount the real one into this fake one